Some events weren't treated as events
Some events weren't treated as events because of how they appear in the d.ts
file. For example, a function declared like this is treated like an event:
  onclick(): void;

but an event declared like this is treated as a property rather than a function
so it isn't recapitalized:
  onclick: Function;

The fix is to check if each function and property is an event (rather than
just function).

Also, properties ending in "element" are no longer ignored. Instead, it's up
to consmuers of the output to ignore such properties that they don't care about.